reset_list.sh udpate
Improve reset_list.sh to solve the MainUI bugs
This script is used as workaround to a MainUI bug: When you have less than 6 roms, you add a rom and delete the current database: MainUI will not update the number of roms in the list until you scroll to the bottom.
Update: modify state.json too when required

(Fix bug when showing/hiding Video - TV channels)

+++ System/usr/trimui/scripts/reset_list.sh
@@ -9,9 +9,11 @@ if [ -z "$1" ]; then
 fi
 
 InputPath="$1"
+
+############################### update romwinidx.json ###############################
+
 JsonFile="/mnt/romwinidx.json"
 
-# Determine if InputPath is a directory name or full path
 if [ -d "$InputPath" ]; then
     RomDir="$InputPath"
     RomDirBase="$(basename "$RomDir")"
@@ -20,15 +22,9 @@ else
     RomDirBase=$InputPath
 fi
 
-# Check if the JSON file exists
-if [ ! -f "$JsonFile" ]; then
-    echo "Error: The file $JsonFile does not exist."
-    exit 1
-fi
-
-# Modify the JSON file using jq
-tmpfile=$(mktemp)
-jq --arg RomDirBase "$RomDirBase" '
+if [ -f "$JsonFile" ]; then
+    tmpfile=$(mktemp)
+    jq --arg RomDirBase "$RomDirBase" '
     .list |= map(
         if (.rompath | contains($RomDirBase)) then
             .end = (.start + 5)
@@ -38,17 +34,45 @@ jq --arg RomDirBase "$RomDirBase" '
     )
 ' "$JsonFile" >"$tmpfile"
 
-# Check if jq succeeded
-if [ $? -eq 0 ]; then
-    mv "$tmpfile" "$JsonFile"
-    echo "Successfully updated rompaths containing \"$RomDir\" or \"$RomDirBase\"."
+    if [ $? -eq 0 ]; then
+        mv "$tmpfile" "$JsonFile"
+        echo "Successfully updated romwinidx.json containing \"$RomDirBase\"."
+    else
+        echo "Error: Failed to update the JSON file."
+        rm -f "$tmpfile"
+    fi
+
 else
-    echo "Error: Failed to update the JSON file."
-    rm -f "$tmpfile"
-    exit 1
+    echo "Error: The file $JsonFile does not exist."
 fi
 
-# Remove cache file if it exists
+############################### update state.json ###############################
+
+stateFile="/tmp/state.json"
+if [ -f "$stateFile" ]; then
+    tmpfile=$(mktemp)
+
+    jq --arg InputPath "$InputPath" '
+        .list |= map(
+            if (.emulaunch? and (.emulaunch | type == "string") and (.emulaunch | contains($InputPath))) then
+                .pageend = (.pagestart + 5)
+            else
+                .
+            end
+        )
+    ' "$stateFile" >"$tmpfile"
+
+    if [ $? -eq 0 ]; then
+        mv "$tmpfile" "$stateFile"
+        echo "Successfully updated state.json."
+    else
+        echo "Error: Failed to update state.json."
+        rm -f "$tmpfile"
+    fi
+fi
+
+############################### Remove cache file if it exists ###############################
+
 CacheFile="$RomDir/${RomDirBase}_cache7.db"
 if [ -f "$CacheFile" ]; then
     rm -f "$CacheFile"
